requirements.txt
fastapi[standard]>=0.111.0
uvicorn[standard]>=0.29.0
aiortc>=1.9.0
openai>=1.30.0
deepgram-sdk>=6.0.0
elevenlabs>=1.3.0
numpy>=1.26.0
scipy>=1.13.0
python-dotenv>=1.0.0
websockets>=12.0
av>=12.0.0
